requirements.txt
PyOpenGL
opencv-python
mediapipe
